Added duo plotter funcitonality

+++ playground/numpy/multid/main.py
@@ -1,5 +1,5 @@
 import numpy as np
 import multid
 
-arr2D = np.array([[1, 1, 1], [2, 2, 3]])
+arr2D = np.array([[1, 1, 1], [2, 2, 3]], dtype=np.int32)
 print multid.multid(arr2D)
